Migrating deprecated_api_test to ginkgo
